# src/Esquio.UI.Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/products/{productName}/features/{featureName}": {
      "put": {
        "tags": [
          "Features"
        ],
        "operationId": "Features_Update",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "featureName",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "updateFeatureRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateFeatureRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Features"
        ],
        "operationId": "Features_Delete",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "featureName",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Features"
        ],
        "operationId": "Features_Get",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "featureName",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DetailsFeatureResponse"
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_List",
        "parameters": [
          {
            "type": "integer",
            "name": "pageIndex",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "pageCount",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListProductResponse"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_Add",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddProductRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Products/{productName}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_Get",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DetailsProductResponse"
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_Update",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProductRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_Delete",
        "parameters": [
          {
            "type": "string",
            "name": "productName",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": {}
        }
      }
    },
...
    "ValidationProblemDetails": {
      "allOf": [
        {
          "$ref": "#/definitions/ProblemDetails"
        },
        {
          "type": "object",
          "properties": {
            "errors": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      ]
    },
...
    "UpdateFeatureRequest": {
      "type": "object",
      "required": [
        "enabled"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        }
      }
    },
    "DetailsFeatureResponse": {
      "type": "object",
      "required": [
        "enabled"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "toggles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ToggleDetail"
          }
        }
      }
    },
    "ToggleDetail": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "friendlyName": {
          "type": "string"
        },
        "parameters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterDetail"
          }
        }
      }
    },
    "ParameterDetail": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
...
    "ListProductResponse": {
      "type": "object",
      "required": [
        "total",
        "pageIndex",
        "count"
      ],
      "properties": {
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "pageIndex": {
          "type": "integer",
          "format": "int32"
        },
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListProductResponseDetail"
          }
        }
      }
    },
    "ListProductResponseDetail": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "DetailsProductResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "AddProductRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "defaultRingName": {
          "type": "string"
        }
      }
    },
...
    "UpdateProductRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
  }
}